refactor: split filter.go out of engine.go and runmain_test.go out of main_test.go
Both files crossed qlty's per-file complexity threshold (50). Moved:
- applyFilterRecursive + applyNotFilter + isSimpleFilter + applySimple-
  Filter + applyGroupFilter + applyGroupStep + extractGroupStep +
  combineBool + FilterDynamicData from engine.go → filter.go
- TestStartServer / TestRunMain_* / TestCloseClient_NoopWhenNil /
  TestMain_ExitCode from main_test.go → runmain_test.go

qlty smells now reports zero issues. No behaviour change.
